Registry / CLI Updates (#1)
* Updates for bpm-cli and public registry

* Add ConfigFilesRenders - utility functions for rendering multiple files

* Allow passing commands as file when starting a docker container

The idea is that plugins create the docker command as part of the
configuration.

This allows overwriting the command before starting the container
for troubleshooting or other tinkering purposes. It also simplifies the
plugin code that starts the container and makes it easier to understand
what's going on "under-the-hood".

* Rename `remove` to `stop`.

This describes the functionality better, as by default we don't really
remove anything that cannot be easily started again

* Revert change that removed expanding the baseDir

The default uses ~ to denote the home directory. Without this, the
plugin isn't usable with the default --baseDir

* Use correct default for `--baseDir`

* Add a status command

* Make isContainerRunning public

This is useful for plugins to determine if all dependencies are running