Add a plugin implementation that provides decent defaults
This makes plugins even simpler. They just have to use this specific
plugin and pass in the configuration/containers.

Note that, in order to do this, I had to change Plugin from a struct
to an interface.

This also paves the path for plugins that target other platforms (e.g.
k8s)