feat: use Go templates in docker mount definitions
This can be used to access parameters (declaratively) to determine what
to mount.

The first use-case for this is to use the `data-dir` parameter to be
able to specify where blockchain data is stored:

    {
	  Type: "bind",
      From: "{{ index .Node.StrParameters \"data-dir\" }}",
      To:   "/data",
	}

In the future we can use the same technique to parameterize even more.

BREAKING CHANGE: changes the interface of docker.go. For most plugins
this should be fine because they use the higher-level docker_plugin.go